requirement.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.3.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
decorator==5.1.1
entrypoints==0.4
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.51.3
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.22.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.4.3
pandas-stubs==1.2.0.56
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
streamlit==1.11.1
tabulate==0.8.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.3.0
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
zipp==3.15.0
